module2/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
pydantic-settings==2.2.1
transformers==4.41.0
sentencepiece==0.2.0
sentence-transformers==3.0.0
python-dotenv==1.0.1
pdfplumber>=0.11.0
python-multipart>=0.0.9
python-docx>=1.1.0
